pyproject.toml
[project]
name = "dbome"
version = "0.2.0"
description = "dbome (dbt at home) - A dbt-like tool for BigQuery view management"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "jinja2>=3.1.0",
    "google-cloud-bigquery>=3.13.0",
    "pyyaml>=6.0.1",
    "rich>=13.0.0",
    "sqlglot[rs]>=18.0.0",
    "boto3>=1.26.0",
    "google-auth>=2.0.0",
    "pydantic>=2.10.6",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
dbome = "dbome.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dbome"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
]
